pyproject.toml
[project]
name = "shieldcraft-ai"
version = "0.1.7"
description = "GenAI-driven cybersecurity platform"
authors = [
    {name = "deon66@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pytest>=8.2.2,<9.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "torch>=2.3.0,<3.0.0",
    "scikit-learn>=1.4.2,<2.0.0",
    "xgboost>=2.0.3,<3.0.0",
    "lightgbm>=4.3.0,<5.0.0",
    "langchain-community==0.3.26",
    "mlflow>=2.12.2,<3.0.0",
    "pandas>=2.2.2,<3.0.0",
    "numpy>=2.1.0,<3.0.0",
    "pyspark>=3.5.1,<4.0.0",
    "kafka-python>=2.0.2,<3.0.0",
    "boto3>=1.34.113,<2.0.0",
    "pgvector>=0.2.4,<1.0.0",
    "neptune-client>=1.8.7,<2.0.0",
    # "sagemaker>=2.224.0,<3.0.0",  # Removed due to numpy 2.x conflict
    "docker>=7.0.0,<8.0.0",
    "kubernetes>=29.0.0,<30.0.0",
    "prometheus-client>=0.20.0,<1.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "requests>=2.32.3,<3.0.0",
    "pydantic>=2.7.1,<3.0.0",
    "aws-cdk-lib>=2.0.0,<3.0.0",
    "constructs>=10.0.0",
    "diagrams>=0.23.4,<1.0.0",
    "graphviz>=0.20.3,<1.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
ruff = "^0.4.4"
black = "^24.4.2"

[tool.poetry.group.typecheck.dependencies]
mypy = "^1.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^4.1.0"
nbval = ">=0.11.0,<0.12.0"
nbqa = "^1.8.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
nox = "2023.4.22"
toml = "^0.10.2"
yamllint = "^1.32.0"
nbval = "^0.11.0"

[tool.poetry.group.security.dependencies]
safety = "^3.2.0"
bandit = "^1.7.8"

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false

[tool.ruff]
ignore = ["E402"]

[tool.pytest.ini_options]
markers = [
    "integration: mark a test as an integration test",
    "msk: mark a test as related to MSK"
]
